refactor: realiza regras de negócio no caso de uso
