+++ app/Application/FinishPendingWithdrawals/FinishPendingWithdrawals.php
@@ -3,13 +3,13 @@
 namespace App\Application\FinishPendingWithdrawals;
 
 use App\Domain\Repository\WithdrawalRepository;
-use App\Application\Withdraw\Withdrawer;
+use App\Domain\Service\AsyncWithdrawDispatcher;
 
 class FinishPendingWithdrawals
 {
     public function __construct(
         private WithdrawalRepository $withdrawalRepository,
-        private Withdrawer $withdrawer,
+        private AsyncWithdrawDispatcher $asyncWithdrawDispatcher,
     ) {
     }
 
@@ -18,7 +18,7 @@ class FinishPendingWithdrawals
         $pendingWithdrawals = $this->withdrawalRepository->findPending();
 
         foreach ($pendingWithdrawals as $withdrawal) {
-            $this->withdrawer->execute($withdrawal);
+            $this->asyncWithdrawDispatcher->dispatch($withdrawal);
         }
     }
 }
