+++ app/Infrastructure/Job/Cron/FinishPendingWithdrawsJob.php
@@ -4,24 +4,18 @@ declare(strict_types=1);
 
 namespace App\Infrastructure\Job\Cron;
 
-use App\Domain\Repository\WithdrawalRepository;
-use App\Domain\Service\AsyncWithdrawDispatcher;
+use App\Application\FinishPendingWithdrawals\FinishPendingWithdrawals;
 use Hyperf\AsyncQueue\Job;
 
 class FinishPendingWithdrawsJob extends Job
 {
     public function __construct(
-        private WithdrawalRepository $withdrawalRepository,
-        private AsyncWithdrawDispatcher $asyncWithdrawDispatcher,
+        private FinishPendingWithdrawals $finishPendingWithdrawals,
     ) {  
     }
 
     public function handle(): void
     {
-        $pendingWithdrawals = $this->withdrawalRepository->findPending();
-
-        foreach ($pendingWithdrawals as $withdrawal) {
-            $this->asyncWithdrawDispatcher->dispatch($withdrawal);
-        }
+        $this->finishPendingWithdrawals->execute();
     }
 }
